Added two bounties for The GIMP
If we had been informed earlier about the two bounties for The GIMP, and if the build system would be easier to use on Win32 - the build.exe doesn't even run, even with .NET installed, isn't that ironic - I would have created the html files as well. I hope someone else does this now.