Combine the What Is GNOME text from ../index.wml and just link to the
2007-08-14  Murray Cumming  <murrayc@murrayc.com>

* docbook/C/release-notes.xml: Combine the What Is GNOME 
text from ../index.wml and just link to the About GNOME
page on the website. Remove the Getting Involved section 
and just link to the Join GNOME page on the website.
Add the list of translations here.
This page can now be the main start/2.20 page itself.
* docbook/Makefile_shared.am_fragment: 
* docbook/C/rngettinginvolved.xml: Removed this xml file.

* docbook/C/rnbase.xml: Removed unused file.

* docbook/C/rnadmins.xml:
* docbook/C/rndevelopers.xml:
* docbook/C/rni18n.xml:
* docbook/C/rninstallation.xml:
* docbook/C/rnlookingforward.xml:
* docbook/C/rnusers.xml: Use the XML entity again for gnomeversion 
and lastversion that was introduced after 2.12.

svn path=/trunk/; revision=5344