Mention RTL in the introductory paragraph. Remove the TODO text which
2007-09-13  Murray Cumming  <murrayc@murrayc.com>

* docbook/C/release-notes.xml: Mention RTL in the introductory 
paragraph.
* docbook/C/rndevelopers.xml: Remove the TODO text which should 
link to the GTK+ announcement because it has not happened yet. 
We can add this in the en translation.
* docbook/C/rni18n.xml: Remove the text about new languages, 
because we either do not have any, or I can't find out that 
information.
* docbook/C/rnusers.xml: Remove an excessive For Instance.

We are now in string freeze. Typos and small improvements may 
be made in the en translation.

svn path=/trunk/; revision=5432