Remove the warning about it being a draft. This removal does not affect
2007-09-19  Murray Cumming  <murrayc@murrayc-desktop>

* docbook/C/release-notes.xml: Remove the warning about it being a 
draft. This removal does not affect translations.
* docbook/C/rni18n.xml: Remove languages that were not translated.
Change the number of languages from 50 to 48 - these translations 
will need to be updated, but we expected that.
* docbook/en/en.po: Change the list of translations of the release 
notes so that we do not mention the untranslated languages. 
This will only be correct in the English version, otherwise we 
will break the translations too much. This seems OK.

svn path=/trunk/; revision=5502